chore: add tutorial for difflogic generation
