--- docs/DLG/.gitignore
@@ -1,3 +1,0 @@
-data-mnist/
-*.svg
-*.pickle
