--- docs/DLG/demo/.gitignore
@@ -1,3 +1,0 @@
-data-mnist/
-img/difflogic.svg
-sim_build/
