+++ docs/tutorials/difflogic/.gitignore
@@ -1,0 +1,3 @@
+data-mnist/
+img/difflogic.svg
+sim_build/
